chore: loading script and badge element loading
