+++ extension/src/content.ts
@@ -1,11 +1,8 @@
 import type { PlasmoCSConfig } from "plasmo"
-import browser from "webextension-polyfill"
 
-import { ChatMutationObserver } from "~scripting/observer"
+import { appLoader } from "~scripting"
 
 export {}
-
-const CHAT_LIST = ".chat-list--default,.chat-list--other,.chat-list"
 
 export const config: PlasmoCSConfig = {
   matches: [
@@ -23,34 +20,5 @@ export const config: PlasmoCSConfig = {
 }
 
 // @ts-ignore
-
-const appLoader = () => {
-  console.log("TBP: Loading Twitch Better Profile...")
-
-  let hasWelcomeMessageOnBody = document.body.querySelector(
-    'div[data-a-target="chat-welcome-message"]'
-  )
-
-  if (!hasWelcomeMessageOnBody) {
-    console.log("TBP: Welcome message not found, waiting...")
-    return setTimeout(appLoader, 3000)
-  }
-
-  let chatElements = document.querySelector(CHAT_LIST)
-  if (!chatElements) {
-    return setTimeout(appLoader, 3000)
-  }
-
-  let chat = chatElements as Node
-  console.log(chat)
-  console.log("TBP: Loaded! Starting to listen to new messages...")
-
-  let observer = new ChatMutationObserver()
-  observer.start(chat, {
-    childList: true,
-    subtree: true,
-    characterData: true
-  })
-}
 
 appLoader()
